requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
huesdk==1.1
idna==3.3
nanoleafapi==2.1.2
Pillow==9.2.0
python-json-logger==2.0.4
requests==2.28.1
six==1.16.0
sseclient==0.0.27
urllib3==1.26.12
websockets==10.3
